journal app with auth

+++ journal-app/src/Journal/DeleteButton.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types';
+import firebase from 'firebase';
 import db from '../firebase/db';
 import { useHistory, useLocation } from 'react-router';
 
@@ -8,7 +9,16 @@ export default function DeleteButton({ id }) {
     const location = useLocation();
 
     const onDelete = () => {
-        db.collection('journalEntries')
+        const user = firebase.auth().currentUser;
+
+        if (!user) {
+            return; // TODO: Handle this gracefully
+        }
+
+        db
+            .collection('users')
+            .doc(user.uid)
+            .collection('journalEntries')
             .doc(id)
             .delete()
             .then(() => {
